requirements.txt
backports.weakref==1.0.post1
bleach==1.5.0
buildozer==0.34
certifi==2018.4.16
chardet==3.0.4
Cython==0.20
docutils==0.14
funcsigs==1.0.2
h5py==2.7.1
html5lib==0.9999999
idna==2.6
Keras==2.1.3
Kivy==1.9.1
Kivy-Garden==0.1.4
Markdown==2.6.11
mock==2.0.0
numpy==1.14.2
opencv-python==3.4.0.12
pbr==4.0.2
pexpect==4.5.0
Pillow==5.1.0
protobuf==3.5.2.post1
ptyprocess==0.5.2
pygame==1.9.3
Pygments==2.2.0
pytesseract==0.2.0
PyYAML==3.12
requests==2.18.4
scipy==1.0.1
six==1.11.0
tensorflow==1.3.0
tensorflow-tensorboard==0.1.8
urllib3==1.22
virtualenv==15.2.0
Werkzeug==0.14.1
google-api-python-client==1.6.5
google-auth==1.4.1
google-auth-httplib2==0.0.3
